+++ Editor/Editor/Gizmo.cpp
@@ -1,5 +1,6 @@
 #include <EditorStdafx.h>
 
+#include <Editor/EditorColours.h>
 #include <Editor/Gizmo.h>
 #include <Editor/Utils/MousePicking.h>